requirements.txt
altair==5.5.0
annotated-types==0.7.0
asn1crypto==1.5.1
atpublic==5.1
attrs==25.1.0
backports.tarfile==1.2.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudpickle==2.2.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
filelock==3.17.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.43
idna==3.10
importlib_metadata==8.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.6.0
narwhals==1.26.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requirements-parser==0.11.0
rich==13.9.4
rpds-py==0.22.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
snowflake-cli==3.3.0
snowflake-connector-python==3.12.4
snowflake-snowpark-python==1.28.0
snowflake.core==1.0.2
sortedcontainers==2.4.0
streamlit==1.42.0
streamlit-elements==0.1.0
tenacity==9.0.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
typer==0.12.5
types-setuptools==75.8.0.20250210
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.2.3
zipp==3.21.0
streamlit